change galleryListBodyPage to ArtowrkListBodyPage
